+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support serving multiple sites/indices.
 - Support config file.
 - MoMA demo site.
+- Filter on size of documents.
 
 ## [0.1.0] - 2022-09-30